+++ lizard_portal/media/lizard_portal/lizard/model/CollageModel.js
@@ -29,10 +29,12 @@
       },
       afterRequest: function(request, success) {
         if (request.method === 'POST') {
-          if (success) {
-            return Ext.MessageBox.alert('Opslaan gelukt');
-          } else {
-            return Ext.MessageBox.alert('Opslaan mislukt');
+          if (request.params.data.indexOf("is_temp") === -1) {
+            if (success) {
+              return Ext.MessageBox.alert('Opslaan gelukt');
+            } else {
+              return Ext.MessageBox.alert('Opslaan mislukt');
+            }
           }
         }
       }
@@ -42,6 +44,9 @@
         name: 'id',
         mapping: 'id',
         type: 'number'
+      }, {
+        name: 'secret_slug',
+        type: 'string'
       }, {
         name: 'name',
         type: 'string'
@@ -53,6 +58,9 @@
         type: 'auto'
       }, {
         name: 'read_only',
+        type: 'boolean'
+      }, {
+        name: 'is_temp',
         type: 'boolean'
       }, {
         name: 'layers',