+++ lizard_portal/media/lizard_portal/lizard/popup/TimeSeriesGraph.js
@@ -45,7 +45,12 @@
         if (record.data.is_collage_item === true) {
           graph_title = 'Collage popup voor ' + record.data.grouping_hint;
           title = record.data.grouping_hint;
-          bbar = [];
+          bbar = [
+            {
+              text: 'Ga naar collagescherm',
+              handler: function(btn, event) {}
+            }
+          ];
         } else {
           graph_title = 'Grafiek voor ' + record.data.geo_ident + ' ' + record.data.par_ident + ' ' + record.data.mod_ident + ' ' + record.data.stp_ident;
           title = workspaceitem.get('title') + ' - ' + record.data.geo_ident;