+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-portal
 0.58.3 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Added link to collage screen.
 
 
 0.58.2 (2012-06-15)